refactor: booking list and creation a new component with name wrapperFlights
