+++ src/components/booking/index.js
@@ -3,4 +3,5 @@ import Location from './location';
 import Reservation from './reservation';
 import Date from './date';
 import Passengers from './passengers';
-export {ButtonNext, Location, Reservation, Date, Passengers};
+import WrapperFlights from './wrapperFlights';
+export {ButtonNext, Location, Reservation, Date, Passengers, WrapperFlights};
