+++ src/container/booking/bookingList.container.js
@@ -1,53 +1,27 @@
-import React, {useEffect, useState} from 'react';
+import React, {useState} from 'react';
 import {View, FlatList, Text} from 'react-native';
-import {ButtonNext, Reservation} from '../../components/booking';
-import {FontAwesomeIcon} from '@fortawesome/react-native-fontawesome';
-import {faPlus, faPowerOff} from '@fortawesome/free-solid-svg-icons';
+import {WrapperFlights, Reservation} from '../../components/booking';
 import {general} from './styles';
-import {useLogout} from '../../library/hooks';
+import {useLogout, useFirebaseGet} from '../../library/hooks';
 import auth from '@react-native-firebase/auth';
-import firestore from '@react-native-firebase/firestore';
 
 const BookingList = props => {
   const [data, setData] = useState([]);
   const reservations = ({item}) => {
     return <Reservation key={item} {...item} />;
   };
-  const [logout] = useLogout(props);
-
-  useEffect(() => {
-    const unsubscribeListener = firestore()
-      .collection('reservas')
-      .doc(auth().currentUser.uid)
-      .onSnapshot(documentSnapshot => {
-        if (documentSnapshot) {
-          var DATA = documentSnapshot.data().flights;
-          setData(DATA);
-        }
-      });
-    return () => unsubscribeListener();
-  }, []);
+  const [logout] = useLogout(props.navigation.navigate);
+  useFirebaseGet(auth().currentUser.uid, setData, 'reservas', 'flights');
   return (
     <View style={general.generalContainer}>
       <Text style={general.tittle}>My flights</Text>
-      <ButtonNext
-        round
-        name={<FontAwesomeIcon icon={faPowerOff} size={24} color="#FFF" />}
-        functionNext={logout}
-        active
-      />
-      <FlatList
-        data={data}
-        renderItem={reservations}
-        keyExtractor={item => item.id}
-      />
-      <ButtonNext
-        position="absolute"
-        round
-        name={<FontAwesomeIcon icon={faPlus} size={24} color="#FFF" />}
-        functionNext={() => props.navigation.navigate('BookingFrom')}
-        active={true}
-      />
+      <WrapperFlights navigate={props.navigation.navigate} logout={logout}>
+        <FlatList
+          data={data}
+          renderItem={reservations}
+          keyExtractor={item => item.id}
+        />
+      </WrapperFlights>
     </View>
   );
 };
