+++ src/library/methods/index.js
@@ -13,5 +13,4 @@ export {
   isObject,
   isString,
   setBase64,
-
 };
